32 bit out, 64 is in man

+++ util/sieve/highest_prime.h
@@ -11,7 +11,7 @@ public:
         init_sieve();
     }
 
-    uint32_t count() {
+    uint64_t count() {
         int count = 0;
         for (int i = 1; i < primes.size(); i++) {
             count += primes[i] == 0;
@@ -20,23 +20,23 @@ public:
     }
 
     void print() {
-        for (uint32_t i = 0; i < primes.size(); i++) {
+        for (uint64_t i = 0; i < primes.size(); i++) {
             cout << i << " " << primes[i] << endl;
         }
     }
 
-    uint32_t highest_prime(uint32_t n) {
+    uint64_t highest_prime(uint64_t n) {
         return primes[n];
     }
 
-    bool is_prime(uint32_t n) {
+    bool is_prime(uint64_t n) {
         return primes[n] == 0;
     }
 
     // Returns the next prime greater than n
     // returns 0 if next prime isn't within sieve.
-    int next_prime(uint32_t n) {
-        uint32_t start = n;
+    int next_prime(uint64_t n) {
+        uint64_t start = n;
         if (n == 0) return 1;
         if (n == 1) return 2;
         if (start % 2 == 0 && start + 1 < primes.size()) {
@@ -44,7 +44,7 @@ public:
             start += 1;
             if (primes[start] == 0) { return start; }
         }
-        for (uint32_t i = start + 2; i < primes.size(); i += 2) {
+        for (uint64_t i = start + 2; i < primes.size(); i += 2) {
             if (primes[i] == 0) { return i; }
         }
         return 0;
@@ -56,7 +56,7 @@ private:
         primes[1] = 1;
         for (uint32_t i = 2; 2 * i < primes.size(); i++) {
             if (primes[i] == 0) {
-                for (uint32_t j = 2 * i; j < primes.size(); j += i) {
+                for (uint64_t j = 2 * i; j < primes.size(); j += i) {
                     primes[j] = i;
                 }
             }
